requirements.txt
aiohttp>=3.8.0
aiodns>=3.0.0
rich>=13.0.0
requests>=2.28.0
dnspython>=2.3.0
colorama>=0.4.6
fake-useragent>=1.1.0
beautifulsoup4>=4.12.0
python-whois>=0.9.0
mmh3>=4.0.0
reportlab>=4.0